[MIPS] MatrixProcsTest is changed, because it was failing on jb@Actions tablet.
test_scale_nofilter() call Clamp and Repeat routines that returns buffer with size of count/2, but in test there was memcmp check with size of count.

test_rotation_affine_filter() and test_perspective_filter() are also changed because clamp and repeat routines called from these tests,
returns buffer with size of count*2, but in tests, there was memcmp check with size of count.

Change-Id: I3534e0c9869343890e0da5b453918ed642d7d355