Build skia test program in arm mode
The skia library is built in non-thumb mode, but several optimisations
are disabled if the test program is built in thumb mode.
The ARM version of SkFloatToFixed is not as accurate as the portable
version... presumably that is ok.

Signed-off-by: Chris Dearman <chris@mips.com>

(cherry picked from commit d9d5b63cf8fb5894a429c8ef795db33ba3bdc4cb)

Change-Id: Iad6f1ea50fabe276239f690f1f63653c725aee08
Signed-off-by: Douglas Leung <douglas@mips.com>